a raft of fixes and adding support for limine terminal

+++ kernel64/src/printf.c
@@ -8,7 +8,7 @@ static int do_printf(const char *fmt, va_list args, fnptr_t fn, void *ptr)
 	unsigned flags, actual_wd, count, given_wd;
 	unsigned char *where, buf[PR_BUFLEN];
 	unsigned char state, radix;
-	long num;
+	long long num;
 
 	//while (printlock);
 	//lock_spinlock(&printlock);
